chore(synthetics): code refactor and corrections associated with IN_PROGRESS
