+++ internal/output/text.go
@@ -215,6 +215,43 @@ func (o *Output) newTableWriter() table.Writer {
 			SeparateHeader:  true,
 		},
 	})
-
+	t.SetStyle(table.Style{
+		Name: "nr-syn-cli-table",
+		Box:  table.StyleBoxRounded,
+		Color: table.ColorOptions{
+			Header: text.Colors{text.Bold},
+		},
+		Options: table.Options{
+			DrawBorder:      true,
+			SeparateColumns: true,
+			SeparateHeader:  true,
+		},
+	})
 	return t
+}
+
+func PrintResultTable(tableData [][]string) {
+	o := &Output{terminalWidth: 200}
+
+	tw := o.newTableWriter()
+	tw.Style().Name = "nr-syn-cli-table"
+
+	// Add the header
+	tw.AppendHeader(table.Row{"Status", "Monitor Name", "Monitor GUID", "isBlocking"})
+
+	// Add the rows
+	for _, row := range tableData {
+		tw.AppendRow(stringSliceToRow(row))
+	}
+
+	// Render the table
+	tw.Render()
+}
+
+func stringSliceToRow(slice []string) table.Row {
+	row := make(table.Row, len(slice))
+	for i, v := range slice {
+		row[i] = v
+	}
+	return row
 }
